cb/requirements.txt
Flask
flask-cors
requests
beautifulsoup4
openai
python-dotenv
